requirements.txt
torch==2.0.1+cu117
torchvision==0.15.2+cu117
torchaudio==2.0.2+cu117

numpy
matplotlib==3.7.2
opencv-python==4.8.0.76
scipy
tqdm==4.65.0
einops==0.4.1
pytorch-lightning==2.0.9
diffusers==0.20.2
accelerate==0.24.1
deepspeed
omegaconf==2.2.3
pyarrow==12.0.1
Pillow==9.5.0
requests==2.28.2
psutil==5.9.5
wandb==0.15.8
PyYAML==6.0.1
timm==0.6.7
transformers==4.35.0
gradio==3.32.0
setuptools==60.2.0
safetensors==0.3.1
tensorboard==2.11.0
open-clip-torch==2.20.0
onnxruntime-gpu==1.15.1
albumentations
kornia
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
flash-attn
deepspeed